backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
cloudinary==1.41.0
bcrypt==4.2.0
pyjwt==2.9.0
reportlab==4.2.2
tensorflow==2.17.0
numpy==1.26.4
pytest==8.3.3
pytest-flask==1.3.0
pytest-cov==5.0.0
Flask-SQLAlchemy==3.1.1  
Flask-Migrate==4.0.7     
python-dotenv==1.0.1
huggingface-hub==0.26.2
joblib==1.4.2
scikit-learn==1.5.2
requests==2.31.0